LIU-395: Move out log test to separate file.
- Also added the `conftest.py` so we properly load the runtime when starting pytest as a suite (as opposed to running invidually).
